requirements.txt
bidict==0.22.1
blinker==1.7.0
boto3==1.33.11
botocore==1.33.11
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dlib==19.24.2
face-recognition==1.3.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Executor==1.0.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
greenlet==3.0.2
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
pycparser==2.21
python-dateutil==2.8.2
python-engineio==4.8.0
python-socketio==5.10.0
requests==2.31.0
s3transfer==0.8.2
setuptools==69.0.2
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
urllib3==2.0.7
waitress==2.1.2
Werkzeug==3.0.1
wsproto==1.2.0
zope.event==5.0
zope.interface==6.1